[design] overall update
Updated the design to change the memory map to better align with the
standard NeoRV32 set up and support the bootloader in a somewhat custom
way, it will boot from an external ram, i.e boot mode 1, so the memory
can be initialized.

Set additional configuration options in NeoRV32, mainly to support
running coremark and running demos using the standard linker script.
Need to add additional peripherals later but this board won't support
much out of the box.

Removed the original fabric LSRAM and added a read only RAM at
0xF000_0000 -> 0xF0FF_FFFF (16MB) address space but actual ram size is
just enough to fit the bootrom. This could be refined down in the future
to be more exact

Added an AXI to APB convertor at 0xF100_0000 with 16MB slots for APB
peripherals. Could also add some AHB slots by modifying the convertor.

Added bootloader hex file built for RV32I and added functionality in the
main Tcl script to add the bootloader to the bootrom ram as this is only
possible after P&R has been run via Tcl afaik.