[software] softconsole workspace fix
SoftConsole was creating a gitignore entry for ./.metadata in some cases
which resulted in no workspace files being saved in the git commit.

Adding base workspace files as changes to these can be ignored unless
the base project is being re-configured. Watch that the gitignore isn't
updated in future inadvertantly removing the workspace